Fix - prevent calling Moodle's delete event after item was already deleted.
I have noticed that, some times, (on different course format. ie, topcoll) the delete event is propagated to Moodle's delete event after the item was already removed from the DOM and an error message is displayed to the user after the item was already successfully removed from the page and database.

Users (Teachers) get the wrong impression and think the item was not deleted. but it was.

Adding e.stopPropagation(); to the end of the event function stops this behaviour.